Cleanup of process dependencies
