+++ ui/server/src/main/scala/pl/touk/nussknacker/ui/api/ServiceRoutes.scala
@@ -2,7 +2,7 @@ package pl.touk.nussknacker.ui.api
 
 import akka.http.scaladsl.marshalling.Marshal
 import akka.http.scaladsl.model.StatusCodes._
-import akka.http.scaladsl.model.{ContentTypes, HttpEntity, HttpResponse, MessageEntity, ResponseEntity}
+import akka.http.scaladsl.model.{HttpResponse, ResponseEntity}
 import akka.http.scaladsl.server.{Directives, ExceptionHandler, Route}
 import com.typesafe.scalalogging.LazyLogging
 import pl.touk.nussknacker.engine.ModelData
@@ -15,11 +15,8 @@ import pl.touk.nussknacker.ui.security.api.{LoggedUser, Permission}
 import scala.concurrent.{ExecutionContext, Future}
 import scala.util.control.NonFatal
 import de.heikoseeberger.akkahttpcirce.FailFastCirceSupport
-import io.circe.{Encoder, Json, ObjectEncoder}
+import io.circe.{Encoder, Json}
 import io.circe.generic.JsonCodec
-import io.circe.syntax._
-import pl.touk.nussknacker.engine.api.process.WithCategories
-import pl.touk.nussknacker.engine.api.test.InvocationCollectors.QueryServiceResult
 import pl.touk.nussknacker.engine.util.json.BestEffortJsonEncoder
 
 class ServiceRoutes(modelDataMap: Map[ProcessingType, ModelData])
@@ -73,20 +70,14 @@ class ServiceRoutes(modelDataMap: Map[ProcessingType, ModelData])
 
   private[api] def canUserInvokeService(user: LoggedUser, serviceName: String, modelData: ModelData): Boolean = {
 
-    def hasUserDeployPermissionForServiceWithCategories(withCategories: WithCategories[_]) = {
-      def isAllowed(categoryName: String): Boolean =
-        user.can(categoryName, Permission.Deploy)
-
-      withCategories.categories.exists(isAllowed)
+    def hasUserDeployPermissionForCategories(categories: List[String]) = {
+      def isAllowed(categoryName: String): Boolean = user.can(categoryName, Permission.Deploy)
+      categories.exists(isAllowed)
     }
 
-    val servicesMap = //TODO: partly dulicated with ServiceQuery.serviceMethodMap
-      modelData.withThisAsContextClassLoader{
-        modelData.configCreator.services(modelData.processConfig)
-      }
+    val servicesToCategories = modelData.processDefinition.services.mapValues(_.categories)
 
-    servicesMap.get(serviceName)
-      .forall(hasUserDeployPermissionForServiceWithCategories)
+    servicesToCategories.get(serviceName).forall(hasUserDeployPermissionForCategories)
   }
 
   private def invokeService(serviceName: String, modelData: ModelData, params: List[Parameter]): Future[QueryResult] = {
